pyproject.toml
[tool.pdm]

[project]
name = "config"
version = "0.1.0"
description = ""
authors = [
    {name = "Camila", email = "camilacristina328santoss@gmail.com"},
]
dependencies = [
    "django>=4.2.1",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.1.0",
    "djangorestframework-simplejwt>=5.2.2",
    "setuptools>=68.1.2",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
